+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Handle old assignments of nonexisting credentials (#79, PLUM Sprint 220715)
 
 ### Features
-- Client management (#77, PLUM Sprint 220520)
+- OpenID Connect client registration (#77, PLUM Sprint 220520)
 
 ### Refactoring
 - Provisioning config in a dedicated JSON file (#80, PLUM Sprint 220729)