+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Release candidate
 
+### Breaking changes
+- Userinfo endpoint no longer accepts `tenant` parameter (#69, PLUM Sprint 220909)
+- Userinfo `resources` is now an object with tenant keys  (#69, PLUM Sprint 220909)
+
 ### Fix
 - Handle old assignments of nonexisting credentials (#79, PLUM Sprint 220715)
 
@@ -11,6 +15,8 @@
 
 ### Refactoring
 - Provisioning config in a dedicated JSON file (#80, PLUM Sprint 220729)
+- Userinfo endpoint no longer accepts `tenant` parameter (#69, PLUM Sprint 220909)
+- Userinfo `resources` is now an object with tenant keys  (#69, PLUM Sprint 220909)
 
 ---